Move player list and resolver to components/core
