+++ src/components/lineup-on-player-list.ts
@@ -7,9 +7,9 @@ import { repeat } from 'lit/directives/repeat.js';
 import { Formation, FormationLine, Position } from '../models/formation.js';
 import { LivePlayer } from '../models/live.js';
 import { PlayerStatus } from '../models/player.js';
+import { PlayerListElement } from './core/player-list-element.js';
 import './lineup-player-card.js';
 import { PlayerCardData } from './lineup-player-card.js';
-import { PlayerListElement } from './player-list-element.js';
 import { SharedStyles } from './shared-styles.js';
 
 interface PlayerLine extends FormationLine {
