+++ src/components/lineup-player-card.ts
@@ -8,9 +8,9 @@ import { classMap } from 'lit/directives/class-map.js';
 import { Position, formatPosition } from '../models/formation.js';
 import { LivePlayer } from '../models/live.js';
 import { PlayerTimeTracker } from '../models/shift.js';
+import { PlayerResolver, playerResolverContext } from './core/player-resolver.js';
 import { synchronizedTimerContext } from './core/synchronized-timer.js';
 import { SynchronizedTimerController } from './core/timer-controller.js';
-import { PlayerResolver, playerResolverContext } from './player-resolver.js';
 import { SharedStyles } from './shared-styles.js';
 
 export interface PlayerCardData {
