+++ test/unit/helpers/mock-player-resolver.ts
@@ -1,6 +1,6 @@
 /** @format */
 
-import { PlayerResolver, playerResolverContext } from '@app/components/player-resolver.js';
+import { PlayerResolver, playerResolverContext } from '@app/components/core/player-resolver.js';
 
 export function mockPlayerResolver(parentNode: HTMLElement, resolver: PlayerResolver) {
   parentNode.addEventListener('context-request', (event) => {
